DeathStarBench/hotelReservation/services-python/profile/requirements.txt
pymongo
python-memcached